+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Fixed
+- Calling `eyes.check` with `fully(false)` now doesn't take full page screenshot even if `forceFullPageScreenshot` is set to true in the configuration. [Trello 1926](https://trello.com/c/4vcerUTm)
+
 ## [3.167.0] - 2020-06-23
 ### Fixed
 - Users can now use fluent API to update configuration in any order. [Trello 1689](https://trello.com/c/UDYmDZnw)